refactor(canvas): update import paths for LiveKit bus and context
Adjusted import paths for `createLiveKitBus` and `CanvasLiveKitContext` in multiple components to ensure correct module resolution. This change enhances code organization and maintains consistency across the project structure.
